+++ javascript/src/index.js
@@ -1,6 +1,7 @@
 require("@babel/polyfill");
 global.h3 = require("h3-js");
 global.vectorize = require("./vectorize.js");
+global.vectorize2 = require("./vectorize2.js");
 global.sayHello = function() {
    console.log("hello");
 };