+++ NEWS.md
@@ -1,3 +1,8 @@
+# h3forr 0.1.3
+
+* Added `manhattan` dataset
+* Added _edge_ functions
+
 # h3forr 0.1.2
 
 * Added functions `h3_to_parent` and `h3_to_children`